Module assembly_fdb_core::file

The structures, as they are serialized

This module contains the low-level structs that make up the FDB file. These structures are annotated with #[repr(C)] and can be used to read directly from a memory-mapped file on a little-endian machine.

Not all values of these structs are valid for FDB files, but all well-formed FDB-files can be represented by these values. Most importantly, the FDBColumnHeader::column_data_type only has a limited amount of defined values but covers the whole 32 bits.

Structs

ArrayHeader

The basic format of an array reference

FDBBucketHeader

The header of a single bucket.

FDBColumnHeader

The header of a column (field-of-row) definition

FDBFieldData

The type and value of a row field.

FDBHeader

The header of the database file.

FDBRowHeader

The header for a single row

FDBRowHeaderListEntry

One entry of the linked list of references to rows.

FDBTableDataHeader

The header of a table data block

FDBTableDefHeader

The header of a table definition

FDBTableHeader

The header of a table.