Trait assembly_fdb_core::value::ValueMapperMut

pub trait ValueMapperMut<TI, TO>
where
    TI: Context,
    TO: Context,
{
    // Required methods
    fn map_string(&mut self, from: &TI::String) -> TO::String;
    fn map_i64(&mut self, from: &TI::I64) -> TO::I64;
    fn map_xml(&mut self, from: &TI::XML) -> TO::XML;
}

Trait for mapping value from one context to another

This traits allows us to implement a generic Value::map function that works similar to three FnMut closures but can guarantee that only one of them ever borrows Self mutably at the same time.

Required Methods

fn map_string(&mut self, from: &TI::String) -> TO::String

Called when mapping a string

fn map_i64(&mut self, from: &TI::I64) -> TO::I64

Called when mapping an i64

fn map_xml(&mut self, from: &TI::XML) -> TO::XML

Called when mapping an XML value