Trait assembly_fdb_core::value::Context

pub trait Context {
    type String;
    type I64;
    type XML;
}

Type-Parameters to Value

This trait is used to parameterize Value to produce the concrete types that are used elsewhere in this crate.

Required Associated Types

type String

The type that holds a ValueType::String

type I64

The type that holds a ValueType::BigInt

type XML

The type that holds a ValueType::VarChar